+++ frontend/src/components/gallery/constants.js
@@ -1,0 +1,3 @@
+// Shared archetype facet helpers — used by ArchetypesZone + ArchetypeCard.
+export const titleCase = (s) => (s ? String(s).replace(/\b\w/g, (c) => c.toUpperCase()) : s);
+export const facetLabel = (v) => titleCase(String(v).replace(' pitch', '').replace(' accent', ''));
