+++ frontend/src/pages/VoiceGallery.jsx
@@ -4,55 +4,19 @@
 //     facet filters to explore hundreds. (core.archetypes / /archetypes API)
 //   • My Imports — a neutral importer: paste any URL you have the rights to, or
 //     upload a file, trim it, save it. The project ships no celebrity catalog.
-import React, { useState, useMemo, useRef, useEffect } from 'react';
+import React, { useState, useRef, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  Search, Download, Play, Pause, Trash2, X, Loader, Star,
-  Wand2, UserPlus, Sparkles, RotateCcw, Grid, List, Upload, Scissors, Store, Send, Package,
-} from 'lucide-react';
-import { Button, Input } from '../ui';
-import { useArchetypeCategories, useArchetypes, useGalleryVoices, useCommunityItems } from '../api/hooks';
+import { Sparkles, Store, Upload } from 'lucide-react';
 import { archetypePreviewUrl, useArchetypeAsProfile } from '../api/archetypes';
-import { addCommunityItem, communitySubmitUrl } from '../api/community';
-import { importPersona } from '../api/profiles';
-import { openExternal } from '../api/external';
-import {
-  searchYoutube, downloadYoutubeClip, deleteGalleryVoice,
-  saveVoiceAsProfile, uploadVoiceClip, previewVoiceUrl,
-} from '../api/gallery';
-import AudioTrimmer from '../components/AudioTrimmer';
+import { previewVoiceUrl } from '../api/gallery';
 import { useAppStore } from '../store';
 import { apiUrl } from '../api/client';
 import { isTauri } from '../utils/media';
 import { claimPlayback, stopActivePlayback } from '../utils/playback';
-import { askConfirm } from '../utils/dialog';
-import { ArchetypeAvatar, ArchetypeIcon, AccentFlag, NowPlaying, USE_CASE_COLOR } from '../utils/archetypeIcons';
+import ArchetypesZone from '../components/gallery/ArchetypesZone';
+import CommunityZone from '../components/gallery/CommunityZone';
+import ImportsZone from '../components/gallery/ImportsZone';
 import './VoiceGallery.css';
-
-const BROWSE_PAGE = 60;
-
-// Facet vocabularies — values must match the backend taxonomy tokens exactly.
-const FACETS = {
-  gender: ['male', 'female'],
-  age: ['child', 'teenager', 'young adult', 'middle-aged', 'elderly'],
-  pitch: ['very low pitch', 'low pitch', 'moderate pitch', 'high pitch', 'very high pitch'],
-  accent: [
-    'american accent', 'british accent', 'australian accent', 'canadian accent',
-    'indian accent', 'chinese accent', 'japanese accent', 'korean accent',
-    'portuguese accent', 'russian accent',
-  ],
-  // English + Chinese come from the generated catalog; the rest are curated
-  // multilingual designed voices. Values must match the archetype `language`
-  // field (a languages.json entry) exactly — that drives the backend filter.
-  lang: [
-    'English', 'Chinese', 'Spanish', 'French', 'German', 'Italian',
-    'Portuguese', 'Russian', 'Hindi', 'Japanese', 'Korean',
-  ],
-};
-
-const titleCase = (s) => (s ? String(s).replace(/\b\w/g, (c) => c.toUpperCase()) : s);
-const facetLabel = (v) => titleCase(String(v).replace(' pitch', '').replace(' accent', ''));
-const hasActiveFilters = (f) => Object.values(f).some((v) => v !== null && v !== '');
 
 export default function VoiceGallery() {
   const { t } = useTranslation();
@@ -234,533 +198,6 @@ export default function VoiceGallery() {
           loadingPreviewId={loadingPreviewId}
           onPlayGallery={(v) => playUrl(apiUrl(previewVoiceUrl(v.id)), v.id)}
           flash={flash}
-        />
-      )}
-    </div>
-  );
-}
-
-// ── Archetypes zone ─────────────────────────────────────────────────────────
-function ArchetypesZone({
-  t, filters, setFilter, resetFilters, favorites, toggleFavorite,
-  viewMode, setViewMode, playingId, loadingPreviewId, onPreview, onUse, onDesign,
-}) {
-  const [favOnly, setFavOnly] = useState(false);
-  const [offset, setOffset] = useState(0);
-  useEffect(() => { setOffset(0); }, [filters]);
-
-  const cleanFilters = useMemo(() => {
-    const out = {};
-    Object.entries(filters).forEach(([k, v]) => { if (v !== null && v !== '') out[k] = v; });
-    return out;
-  }, [filters]);
-
-  // The Featured strip shows only when nothing is filtered; in that case Browse
-  // excludes featured to avoid duplicating it. Once any filter is active the
-  // Featured strip is hidden (see below), so Browse must include featured too —
-  // otherwise the curated multilingual languages (Spanish/French/…), which have
-  // *only* featured archetypes, would filter down to an empty list.
-  const showFeatured = !hasActiveFilters(filters) && !favOnly;
-
-  const categoriesQ = useArchetypeCategories();
-  const featuredQ = useArchetypes({ featured: true, limit: 100 });
-  const browseQ = useArchetypes({
-    ...cleanFilters,
-    ...(showFeatured ? { featured: false } : {}),
-    limit: BROWSE_PAGE,
-    offset,
-  });
-
-  const categories = categoriesQ.data || [];
-  const featured = featuredQ.data?.items || [];
-  const browse = browseQ.data?.items || [];
-  const total = browseQ.data?.total ?? 0;
-
-  const favSet = useMemo(() => new Set(favorites), [favorites]);
-  const applyFav = (list) => (favOnly ? list.filter((a) => favSet.has(a.id)) : list);
-
-  // NOTE: no `key` here — React keys must be passed directly on the element,
-  // not spread in (spreading a `key` prop triggers a dev warning + is ignored).
-  const cardProps = (a) => ({
-    a, t, viewMode,
-    isFavorite: favSet.has(a.id),
-    isPlaying: playingId === a.id,
-    isLoadingPreview: loadingPreviewId === a.id,
-    onPreview, onUse, onDesign, onToggleFavorite: toggleFavorite,
-  });
-
-  return (
-    <div className="gallery-content gallery-scroll">
-      <div className="facet-bar">
-        {/* Three filter lanes (categories · facets · toggles), each its own
-            horizontally-scrollable portion; the view toggle is pinned right. */}
-        <div className="facet-group facet-group--cats use-case-chips">
-          <button className={`category-chip ${!filters.use_case ? 'selected' : ''}`} onClick={() => setFilter('use_case', null)}>
-            {t('gallery.all', { defaultValue: 'All' })}
-          </button>
-          {categories.map((c) => (
-            <button
-              key={c.id}
-              className={`category-chip ${filters.use_case === c.id ? 'selected' : ''}`}
-              onClick={() => setFilter('use_case', filters.use_case === c.id ? null : c.id)}
-              title={c.name}
-            >
-              <ArchetypeIcon name={c.icon} size={13} />
-              {t(`archetypes.use_${c.id}`, { defaultValue: c.name })}
-            </button>
-          ))}
-        </div>
-
-        <div className="facet-group facet-group--facets facet-selects">
-          {['gender', 'age', 'pitch', 'accent', 'lang'].map((dim) => (
-            <select
-              key={dim}
-              className="facet-select"
-              value={filters[dim] ?? ''}
-              onChange={(e) => setFilter(dim, e.target.value || null)}
-            >
-              <option value="">{t(`archetypes.facet_${dim}`, { defaultValue: titleCase(dim) })}</option>
-              {FACETS[dim].map((opt) => <option key={opt} value={opt}>{facetLabel(opt)}</option>)}
-            </select>
-          ))}
-        </div>
-
-        <div className="facet-group facet-group--toggles">
-          <label className="facet-toggle">
-            <input
-              type="checkbox"
-              checked={filters.whisper === true}
-              onChange={(e) => setFilter('whisper', e.target.checked ? true : null)}
-            />
-            {t('archetypes.facet_whisper', { defaultValue: 'Whisper' })}
-          </label>
-          <label className="facet-toggle">
-            <input type="checkbox" checked={favOnly} onChange={(e) => setFavOnly(e.target.checked)} />
-            <Star size={12} /> {t('gallery.favorites', { defaultValue: 'Favorites' })}
-          </label>
-          <button className="facet-reset" onClick={() => { resetFilters(); setFavOnly(false); }}>
-            <RotateCcw size={12} /> {t('gallery.reset', { defaultValue: 'Reset' })}
-          </button>
-        </div>
-
-        <div className="view-toggle">
-          <button className={viewMode === 'grid' ? 'active' : ''} onClick={() => setViewMode('grid')} title="Grid"><Grid size={14} /></button>
-          <button className={viewMode === 'list' ? 'active' : ''} onClick={() => setViewMode('list')} title="List"><List size={14} /></button>
-        </div>
-      </div>
-
-      {showFeatured && (
-        <section className="archetype-section">
-          <div className="content-header"><div className="content-title">{t('archetypes.featured', { defaultValue: 'Featured' })}</div></div>
-          <div className={`archetype-grid ${viewMode}`}>
-            {applyFav(featured).map((a) => <ArchetypeCard key={a.id} {...cardProps(a)} />)}
-          </div>
-        </section>
-      )}
-
-      <section className="archetype-section">
-        <div className="content-header">
-          <div className="content-title">
-            {t('archetypes.browse_all', { defaultValue: 'Browse all' })}
-            <span className="count-badge">{total}</span>
-          </div>
-        </div>
-        {browseQ.isLoading ? (
-          <div className="loading"><Loader className="spin" size={18} /></div>
-        ) : (
-          <>
-            <div className={`archetype-grid ${viewMode}`}>
-              {applyFav(browse).map((a) => <ArchetypeCard key={a.id} {...cardProps(a)} />)}
-            </div>
-            {applyFav(browse).length === 0 && (
-              <div className="empty">{t('gallery.no_matches', { defaultValue: 'No voices match these filters.' })}</div>
-            )}
-            {offset + BROWSE_PAGE < total && !favOnly && (
-              <div className="load-more">
-                <Button variant="ghost" onClick={() => setOffset(offset + BROWSE_PAGE)} disabled={browseQ.isFetching}>
-                  {browseQ.isFetching ? <Loader className="spin" size={14} /> : null}
-                  {t('gallery.load_more', { defaultValue: 'Load more' })}
-                </Button>
-              </div>
-            )}
-          </>
-        )}
-      </section>
-    </div>
-  );
-}
-
-// ── Archetype card ───────────────────────────────────────────────────────────
-function ArchetypeCard({
-  a, t, viewMode, isFavorite, isPlaying, isLoadingPreview,
-  onPreview, onUse, onDesign, onToggleFavorite,
-}) {
-  const color = USE_CASE_COLOR[a.use_case] || '#83a598';
-  const sub = [a.facets.gender, a.facets.age, a.facets.pitch]
-    .filter(Boolean).map(facetLabel).join(' · ');
-  const dialect = a.attrs?.ChineseDialect && a.attrs.ChineseDialect !== 'Auto' ? a.attrs.ChineseDialect : null;
-  const accentLabel = a.facets.accent
-    ? facetLabel(a.facets.accent)
-    : (dialect || (a.language === 'Chinese' ? 'Chinese' : null));
-
-  return (
-    <div className={`archetype-card ${viewMode} ${isPlaying ? 'playing' : ''}`} style={{ '--card-accent': color }}>
-      <div className="arch-head">
-        <ArchetypeAvatar item={a} />
-        <div className="arch-title">
-          <div className="archetype-name">{a.name}</div>
-          {sub && <div className="archetype-sub">{sub}</div>}
-        </div>
-        <button
-          className={`fav-btn ${isFavorite ? 'on' : ''}`}
-          onClick={() => onToggleFavorite(a.id)}
-          title={t('gallery.favorite', { defaultValue: 'Favorite' })}
-        >
-          <Star size={15} fill={isFavorite ? 'currentColor' : 'none'} />
-        </button>
-      </div>
-
-      {/* Always render the chip row (even when empty) so every card shares the
-          same height and the action rows align across the grid. */}
-      <div className="archetype-chips">
-        {accentLabel && (
-          <span className="facet-chip with-flag">
-            <AccentFlag accent={a.facets.accent} lang={a.language} size={14} />
-            {accentLabel}
-          </span>
-        )}
-        {a.facets.whisper && (
-          <span className="facet-chip">{t('archetypes.facet_whisper', { defaultValue: 'Whisper' })}</span>
-        )}
-      </div>
-
-      <div className="arch-foot">
-        <button className="preview-btn" onClick={() => onPreview(a)} title={t('gallery.preview', { defaultValue: 'Preview' })}>
-          {isLoadingPreview ? <Loader className="spin" size={15} /> : isPlaying ? <NowPlaying color={color} /> : <Play size={15} />}
-          <span>{t('gallery.preview', { defaultValue: 'Preview' })}</span>
-        </button>
-        <button className="use-btn" onClick={() => onUse(a)}>
-          <UserPlus size={14} /> {t('gallery.use_voice', { defaultValue: 'Use voice' })}
-        </button>
-        <button className="designer-btn" onClick={() => onDesign(a)} title={t('gallery.open_designer', { defaultValue: 'Open in Designer' })}>
-          <Wand2 size={14} />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// ── Community zone (marketplace) ─────────────────────────────────────────────
-function CommunityZone({ t, playingId, loadingPreviewId, favorites, toggleFavorite, onPlayAudio, flash, onDesign }) {
-  const itemsQ = useCommunityItems({ limit: 100 });
-  const items = itemsQ.data?.items || [];
-  const favSet = useMemo(() => new Set(favorites), [favorites]);
-
-  const submit = async (type) => {
-    try {
-      const { url } = await communitySubmitUrl(type);
-      await openExternal(url);
-    } catch {
-      flash(t('gallery.submit_failed', { defaultValue: 'Could not open the submission form.' }));
-    }
-  };
-
-  return (
-    <div className="gallery-content gallery-scroll">
-      <div className="import-explainer community-explainer">
-        <span>{t('gallery.community_explainer', { defaultValue: 'Designed presets and recorded voices shared by the community, loaded from the omnivoice-gallery.' })}</span>
-        <div className="submit-actions">
-          <button className="submit-btn" onClick={() => submit('preset')}>
-            <Send size={13} /> {t('gallery.submit_preset', { defaultValue: 'Submit a preset' })}
-          </button>
-          <button className="submit-btn" onClick={() => submit('voice')}>
-            <Send size={13} /> {t('gallery.submit_voice', { defaultValue: 'Submit a voice' })}
-          </button>
-        </div>
-      </div>
-
-      {itemsQ.isLoading ? (
-        <div className="loading"><Loader className="spin" size={18} /></div>
-      ) : items.length === 0 ? (
-        <div className="empty">
-          {t('gallery.community_empty', { defaultValue: 'No community voices loaded yet — connect to the internet and reopen, or be the first to submit one.' })}
-        </div>
-      ) : (
-        <div className="archetype-grid grid">
-          {items.map((it) => (
-            <ArchetypeCard
-              key={it.id}
-              a={it}
-              t={t}
-              viewMode="grid"
-              isFavorite={favSet.has(it.id)}
-              isPlaying={playingId === it.id}
-              isLoadingPreview={loadingPreviewId === it.id}
-              onToggleFavorite={toggleFavorite}
-              onPreview={(item) => (item.audio?.url
-                ? onPlayAudio(item.audio.url, item.id)
-                : flash(t('gallery.no_preview', { defaultValue: 'No preview — add it with "Use voice" to hear it.' })))}
-              onUse={async (item) => {
-                try {
-                  const r = await addCommunityItem(item.id, item.name);
-                  flash(t('gallery.saved_as_profile', { defaultValue: 'Added "{{name}}" to your voices.', name: r.name }));
-                } catch {
-                  flash(t('gallery.use_failed', { defaultValue: 'Could not add that voice.' }));
-                }
-              }}
-              onDesign={(item) => (item.instruct
-                ? onDesign(item.instruct)
-                : flash(t('gallery.no_designer', { defaultValue: 'Recorded voice — use "Use voice" instead.' })))}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ── My Imports zone (neutral importer) ───────────────────────────────────────
-function ImportsZone({ t, playingId, loadingPreviewId, onPlayGallery, flash }) {
-  const [query, setQuery] = useState('');
-  const [results, setResults] = useState([]);
-  const [isSearching, setIsSearching] = useState(false);
-  const [isDownloading, setIsDownloading] = useState(false);
-  const [trimming, setTrimming] = useState(null); // { voice, file }
-  const fileRef = useRef(null);
-  const personaRef = useRef(null);
-  const [importingPersona, setImportingPersona] = useState(false);
-
-  const voicesQ = useGalleryVoices();
-  const voices = voicesQ.data || [];
-  const reload = () => voicesQ.refetch();
-
-  // Import a portable .ovsvoice (or legacy .omnivoice) persona bundle (#29).
-  const handlePersonaImport = async (e) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setImportingPersona(true);
-    try {
-      const fd = new FormData();
-      fd.append('file', file);
-      const res = await importPersona(fd);
-      reload();
-      flash(t('gallery.persona_imported', {
-        defaultValue: 'Imported "{{name}}"{{unverified}}.', name: res.name,
-        unverified: res.verified_own_voice ? '' : t('gallery.persona_unverified_suffix', { defaultValue: ' (unverified)' }),
-      }));
-    } catch (err) {
-      const code = String(err?.message || err);
-      const msg = code.includes('413')
-        ? t('gallery.persona_too_large', { defaultValue: 'That bundle is too large (max 100 MB).' })
-        : t('gallery.persona_import_failed', { defaultValue: 'Could not import that persona bundle.' });
-      flash(msg);
-    } finally {
-      setImportingPersona(false);
-      if (personaRef.current) personaRef.current.value = '';
-    }
-  };
-
-  const isUrl = /^https?:\/\//i.test(query.trim());
-
-  const handleSearch = async () => {
-    const q = query.trim();
-    if (!q) return;
-    if (isUrl) {
-      setIsDownloading(true);
-      try {
-        await downloadYoutubeClip({
-          video_url: q, start_time: 0, duration: 15,
-          character_name: t('gallery.imported_clip', { defaultValue: 'Imported clip' }),
-          category: 'import', description: q,
-        });
-        reload();
-        setQuery('');
-      } catch (e) {
-        flash(t('gallery.download_failed', { defaultValue: 'Download failed: {{msg}}', msg: e.message }));
-      } finally {
-        setIsDownloading(false);
-      }
-      return;
-    }
-    setIsSearching(true);
-    try {
-      const r = await searchYoutube(q, 'import', 10);
-      setResults(r.results || []);
-    } catch (e) {
-      flash(t('gallery.search_failed', { defaultValue: 'Search failed.' }));
-    } finally {
-      setIsSearching(false);
-    }
-  };
-
-  const handleDownload = async (info) => {
-    setIsDownloading(true);
-    try {
-      await downloadYoutubeClip({
-        video_url: `https://youtube.com/watch?v=${info.video_id}`,
-        start_time: 0,
-        duration: Math.min(parseFloat(info.duration) || 15, 30),
-        character_name: (info.title || '').substring(0, 40),
-        category: 'import',
-        description: info.title,
-      });
-      reload();
-      setResults([]);
-    } catch (e) {
-      flash(t('gallery.download_failed', { defaultValue: 'Download failed: {{msg}}', msg: e.message }));
-    } finally {
-      setIsDownloading(false);
-    }
-  };
-
-  const handleUpload = async (e) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    const fd = new FormData();
-    fd.append('name', file.name.replace(/\.[^.]+$/, ''));
-    fd.append('category', 'import');
-    fd.append('audio', file);
-    try {
-      await uploadVoiceClip(fd);
-      reload();
-    } catch (err) {
-      flash(t('gallery.upload_failed', { defaultValue: 'Upload failed.' }));
-    } finally {
-      if (fileRef.current) fileRef.current.value = '';
-    }
-  };
-
-  const handleSaveProfile = async (v) => {
-    try {
-      await saveVoiceAsProfile(v.id, v.name);
-      flash(t('gallery.saved_as_profile', { defaultValue: 'Added "{{name}}" to your voices.', name: v.name }));
-    } catch (e) {
-      flash(t('gallery.save_failed', { defaultValue: 'Could not save profile.' }));
-    }
-  };
-
-  const handleDelete = async (v) => {
-    if (!(await askConfirm(t('gallery.confirm_delete', { defaultValue: 'Delete "{{name}}"?', name: v.name })))) return;
-    try { await deleteGalleryVoice(v.id); reload(); } catch { /* noop */ }
-  };
-
-  const handleTrimClick = async (v) => {
-    try {
-      const resp = await fetch(apiUrl(previewVoiceUrl(v.id)));
-      if (!resp.ok) throw new Error('fetch failed');
-      const blob = await resp.blob();
-      const file = new File([blob], `${v.name}.wav`, { type: 'audio/wav' });
-      setTrimming({ voice: v, file });
-    } catch (e) {
-      flash(t('gallery.trim_load_failed', { defaultValue: 'Could not load audio for trimming.' }));
-    }
-  };
-
-  const handleConfirmTrim = async (trimmedFile) => {
-    if (!trimming) return;
-    const { voice } = trimming;
-    const fd = new FormData();
-    fd.append('name', `${voice.name} (Cropped)`);
-    fd.append('character', voice.character || '');
-    fd.append('category', 'import');
-    fd.append('description', voice.description || '');
-    fd.append('audio', trimmedFile);
-    try { await uploadVoiceClip(fd); reload(); setTrimming(null); } catch (e) {
-      flash(t('gallery.upload_failed', { defaultValue: 'Upload failed.' }));
-    }
-  };
-
-  return (
-    <div className="gallery-content">
-      <div className="import-explainer">
-        {t('gallery.import_explainer', {
-          defaultValue: 'Paste a URL you have the rights to (or upload a file), trim the part you need, and save it as a voice. You are responsible for the licensing of anything you import.',
-        })}
-      </div>
-
-      <div className="gallery-search">
-        <div className="search-row">
-          <Input
-            placeholder={t('gallery.import_placeholder', { defaultValue: 'Paste a video/audio URL, or type to search…' })}
-            value={query}
-            onChange={(e) => setQuery(e.target.value)}
-            onKeyDown={(e) => { if (e.key === 'Enter') handleSearch(); }}
-          />
-          <Button onClick={handleSearch} disabled={isSearching || isDownloading} size="sm">
-            {isSearching || isDownloading ? <Loader size={14} className="spin" /> : isUrl ? <Download size={14} /> : <Search size={14} />}
-          </Button>
-          <input ref={fileRef} type="file" accept="audio/*,video/*" hidden onChange={handleUpload} />
-          <Button variant="ghost" size="sm" onClick={() => fileRef.current?.click()} title={t('gallery.upload', { defaultValue: 'Upload file' })}>
-            <Upload size={14} />
-          </Button>
-          <input ref={personaRef} type="file" accept=".ovsvoice,.omnivoice" hidden onChange={handlePersonaImport} />
-          <Button variant="ghost" size="sm" disabled={importingPersona}
-            onClick={() => personaRef.current?.click()}
-            title={t('gallery.import_persona', { defaultValue: 'Import a .ovsvoice persona bundle' })}>
-            {importingPersona ? <Loader size={14} className="spin" /> : <Package size={14} />}
-          </Button>
-        </div>
-      </div>
-
-      {results.length > 0 && (
-        <div className="search-results-panel">
-          <div className="panel-header">
-            <span>{t('gallery.search_results', { defaultValue: '{{count}} results', count: results.length })}</span>
-            <button className="close-btn" onClick={() => setResults([])}><X size={14} /></button>
-          </div>
-          <div className="results-list">
-            {results.map((r, i) => (
-              <div key={i} className="result-row">
-                <div className="result-info">
-                  <span className="result-title">{r.title}</span>
-                  <span className="result-meta">{r.duration || '?'}s</span>
-                </div>
-                <Button size="sm" onClick={() => handleDownload(r)} disabled={isDownloading}>
-                  <Download size={12} /> {t('gallery.import', { defaultValue: 'Import' })}
-                </Button>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      <div className="content-header">
-        <div className="content-title">
-          {t('gallery.my_imports', { defaultValue: 'My Imports' })}<span className="count-badge">{voices.length}</span>
-        </div>
-      </div>
-
-      {voicesQ.isLoading ? (
-        <div className="loading"><Loader className="spin" size={18} /></div>
-      ) : voices.length === 0 ? (
-        <div className="empty">{t('gallery.no_imports', { defaultValue: 'Nothing imported yet. Paste a URL above to get started.' })}</div>
-      ) : (
-        <div className="voice-list">
-          {voices.map((v) => (
-            <div key={v.id} className="voice-card">
-              <button className="voice-play" onClick={() => onPlayGallery(v)}>
-                {loadingPreviewId === v.id ? <Loader className="spin" size={16} /> : playingId === v.id ? <Pause size={16} /> : <Play size={16} />}
-              </button>
-              <div className="voice-info">
-                <span className="voice-name">{v.name}</span>
-                <span className="voice-meta">{Math.round(v.duration || 0)}s</span>
-              </div>
-              <div className="voice-actions">
-                <button className="action-btn" onClick={() => handleTrimClick(v)} title={t('gallery.trim', { defaultValue: 'Trim' })}><Scissors size={14} /></button>
-                <button className="action-btn" onClick={() => handleSaveProfile(v)} title={t('gallery.use_voice', { defaultValue: 'Use voice' })}><UserPlus size={14} /></button>
-                <button className="action-btn danger" onClick={() => handleDelete(v)} title={t('gallery.delete', { defaultValue: 'Delete' })}><Trash2 size={14} /></button>
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-
-      {trimming && (
-        <AudioTrimmer
-          file={trimming.file}
-          maxSeconds={60}
-          onConfirm={handleConfirmTrim}
-          onCancel={() => setTrimming(null)}
         />
       )}
     </div>
